Use native Go implementation of zstd as default
